ops/cfn: bring readyset-postgres CFN templates to a functional state
The goal of this CL is to shore up the last bits of CloudFormation
templates that are otherwise preventing us from deploying a functional
ReadySet for PostgreSQL Cloudformation stack.

While working through deploying the pg stack, I ran into quite a few
errors in Consul server and agents. These errors lead me to review the
security group rules in terms of ingress traffic, and realized we were
missing an allowance for LAN serf gossiping over UDP. I took the
opportunity to really pin down the security group rules for both Consul
server and ec2s that use AMIs w/ consul agents included.

For a list of Consul server/agent ports and their purposes:

https://www.consul.io/docs/install/ports

Adjusted set-host-description.sh to include the AWS instance ID of the
EC2 instance, to ensure that two hosts never have the same hostname.
Through testing and analyzing logs, Consul agents and server weren't
able to communicate correctly due to hostname conflicts. Suffixing the
hostname with the EC2 instance ID is an easy way to prevent theoretical
conflicts going forward.

Additionally, it was discovered that we're missing some certificates
that are required when hitting RDS instances. In either case, it seems
helpful for us to include the requires CA certificates for AWS RDS in
our AMIs, and so I've taken care of that in this CL.

Change-Id: I5abcb2a3d93fb256da20982213c16bd75e046fa7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1947
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>