Remove float precision
It seems that the precision was originally added to emulate
decimal using float and double types, but since we now have
a dedicated decimal type (that doesn't have precision sadly)
we no longer need it.

Moreover it was causing correctness issues, as comparison and
base table lookups were requiring a match on the precision and
not just the enclosed value.

For example, looking up a value by '0' would perform a lookup
with Float(0.0, 255), which is good, but lookup by 0.0 would
perform a lookup by Float(0.0, 1) and fail.

Change-Id: I8088f4d9821053ee3aae09c61d992811eaf6d916
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1951
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>