mir: Escape graphviz output
Exactly the same as we did for dataflow graphviz in 67696bbf8 (Escape <
and > in graphviz output, 2021-07-09), escape < and > in graphviz output
for MIR visualization, so that those characters appearing in filters
don't prevent any of the other content from appearing in the label due
to `dot` attempting to interpret them as HTML tags.

Maybe *someday* we might want to generalize this, but for now it's only
happening twice so I figure it's fine.

Change-Id: I9c226c4e7c4ecfbb27dcc7aceb91ef878960d14a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1952
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>