mir: Add the start of a decorrelation pass
Add the very beginnings of an algebraic decorrelation / unnesting pass
to MIR, which currently has the ability to push simple correlated
filters past basic nodes and below dependent joins, making them no
longer dependent. This now creates the correct graph for a simple
correlated EXISTS query, but actual execution of the query breaks due to
a preexisting materialization-planner bug.

Since this is the second large MIR rewrite pass, this also factors out a
mir::rewrite module, with the existing pull_required_base_columns pass
moved to a submodule of that.

Change-Id: I1ba2453a1bb3eca8682e8a16b80b909e392f61a9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1953
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>