Begin replacing String usages with a SqlIdentifier
The dataflow state is using a whole lot of Strings,
including lots and lots of comparisons and clones.
Moreover many of the strings, used to describe table,
column and node names, are usually very short.

To this end, this CL attempts to replace String usages
with a new SqlIdentifier type, that is very similar to
how we use strings in DataType, specifically inlining
short strings, and using ThinArc for others.

Change-Id: I1bd47777c6f69db43009f367c2126568cf75456f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1965
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>
Tested-by: Buildkite CI