migrate: hoist weak indices alongside strict indices
Lookup obligations in dataflow aren't always directly resolved by the
immediate parent of a join - if that parent is an operator that supports
query_through, the lookup will be done on the *parent's* parent (etc, if
that node in turn also supports query_through), then processed through
the parent. The materialization planner has to be aware of this, since
it needs to "hoist" lookup obligations through any query_through nodes
so that the correct node receives the lookup. This was previously
happening correctly for strict indices, but not for weak indices, which
were added after the fact and special-cased such that they weren't
subject to this hoisting logic. This fixes that, by threading them (as
part of the SuggestedIndex enum) through the `map_indices` function in
the materialization planner. This (coincidentally) fixes the `exists`
logictest when actually fed data, since that happened to create a graph
where a query_through node was a parent of a join, which creates weak
indices for each of its parents.

Change-Id: Ie4b48d74a29aec3e96240627d6acd75f76a94a1c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1957
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>