Type System: Add type to dataflow nodes
Replaces the data flow node column names with a new column struct for
use in dataflow. The new struct contains the column name, the type, and
the source of the column (where it originated).

The entire ColumnSpecification of the returned columns in a query is
required to return the resultset (not just the type). For this reason, a
column source field is maintained so that in the future we can lookup
the specification without tracing the column to where it originates when
we want the specification.

The types for the dataflow nodes are derived in mir_to_flow when the
dataflow node is created. Typically, the new node mirrors the columns of
its parent, except when the column is derived at that new node (when the
node contains a function, expression, or projected literal).

The names of the dataflow columns mirror the names of the associated MIR
columns, rather than the names of the node's parent, to ensure that
aliases are applied at the correct point in dataflow.

Change-Id: I6e4e2a320b6d57708b5813ed3bf8719fa29c0d36
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1960
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>