query_generator: Add support for [un]correlated EXISTS
Add a new Exists variant to query_generator::SubqueryPosition, with a
field specifying whether the generated subquery should be correlated or
not (later we should do the same [un]correlated thing with
SubqueryPosition::Join), and use this to generate EXISTS filters with
subqueries that are either correlated or uncorrelated.

This comes with a generated logictest with *just* exists proving that
this works as expected.

Change-Id: Ifa2b6709ec81669a5f64e7f8ac65ecd6a87fb6a0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1962
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>