Enable full lifecycle tracing for SELECT queries
This covers the query path for queries coming in from the adapter and -
assuming they're Noria-compatible queries - passed down to Noria.  The
adapter will determine whether or not a query should be traced based on
the sample rate and a pseudorandom number; if so, it will pass the trace
ID along to Noria with the rest of the request.

Because the trace IDs are equal, when viewed in a tracing solution (e.g.
Jaeger), spans from both Noria and the adapter representing a single
query appear as one trace.

Change-Id: I68b783595d74645db452e92be3c6ed4b279576cb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2038
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>