CREATE CACHED QUERY: Implement creation by id
Implement CREATE CACHED QUERY [name] AS <QUERY ID>. Allows the user to
get ids for queries not yet migrated with SHOW PROXIED QUERIES and use
those ids to perform a migration.

This is implemented by adding an ids map in the query status cache that
maps a query id (currently made from the query hash) to the
SelectStatement. QueryStatusCache::inner is renamed to
QueryStatusCache::statuses. The statuses and ids maps both reference
Arc<SelectStatement> to avoid duplicating the SelectStatements.
DashMap's default hasher resolves Arc<K> and K to the same hash.

Change-Id: I49530dc00e393cee8a026f4297a6b6f08e1dbc6b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1973
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>