Introduce flag to immediately query fallback on reader cache miss
The upquery path incurs some latency due to potentially executing
through multiple network hops with multiple (de)serialization points.
This CL optimizes the query execution path by instead going immediately
to the upstream database on a cache miss while we asynchronously
perform the upquery - warming the cache for future queries.

This CL expands upon the existing support for non-blocking reads.
Instead of returning whatever results are currently at the reader, we
instead return an explicit signal that a miss occurred. This allows us
to throw an error in the adapter to trigger executing against
fallback. This approach was taken instead of throwing the error in the
server as we perform multiple layers of error wrapping and logging
in the RPC path.

Change-Id: Id4dbdcfcb1f549b90180776b30979685ad52a93f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1980
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>