dataflow: Implement naïve Paginate operator
Implement a first cut at a naïve, fully materialized Paginate
operator, albeit one that fully supports all kinds of reads and writes.
The internals are based on, and almost identical to, the existing
implementation of the topk operaator, except that instead of tracking
whether each row in the current group is new or not, we just use the
presence or absence of a page number column at the end of the row, and
if the page number of a row changes we emit a negative for the old page
number and a positive for the new page number.

This is "naïve" because we don't implement the column_source API for
queries that include the page number, meaning that any replay paths that
include that page number cause the paginate node to be fully
materialized (essentially meaning it will always be fully materialized).
Up next, we'll first implement column_source and upquery remapping to
make Paginate able to at least be partial *by group*, and afterwards put
in the (considerable) effort to allow it to be partial *by page*.

Change-Id: I09359f4b440127e1a9c90180357d45389eb5564f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1981
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>