Fix aarch64 build for cloudflare-zlib
CPU feature detection doesn't currently work on stable on aarch64 for
the cloudflare-zlib library, so this diff adds a flag to override the
feature detection, and force the code to assume the features are always
present on aarch64.

Without this change, noria-server would panic at startup on aarch64
machines (like the M1 MBA).

Change-Id: I1142a8b8c304a13a6fd05200f71bc0c5c3e8e451
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1986
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Tested-by: Buildkite CI