Temporarily introduce the reader layer in the adapter
This CL introduces a noria-server layer into the adapter. This allows
us to perform reads without serialization/deserialization and tcp
overhead. The read path instead reads directly from the reader nodes
in the adapter. This does not include the full subset of flags that
can be passed to a noria-server instance as its intent is only for
proof-of-concepts.

Change-Id: I34baba472cb4f9e9e6abed0817e0dd87a906fb4f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2000
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>