metrics: Add per-query cache miss tracking metric
This adds a new metric, QUERY_LOG_UPQUERIES, which records the total
number of upqueries across all exeuctions of a query. It's logged when
query logging is enabled in the adapter.

Previously we didn't have all the information we needed in the same
process, so this adds an extra stats object that gets returned on reads,
which we can later add more values to and not have to worry about manually
hooking it through the system.

Change-Id: I42ffa205b9f29ae8e5db879f9c26d0bb2f80fdf6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1988
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>