ci: Save release docker images
Made asking for a release vs a debug image now something that must be
explicitly asked for by the build script.

Made it possible to always get a release image from any commit. This
should allow the docker benchmarking to be folded in eventually.

Added a save buildkite upload step at the end for the release steps.

Change-Id: I9cc9c5f91e1897d56e8d31d6787d78e7dd58a0de
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1991
Tested-by: Buildkite CI
Reviewed-by: Ryan Cerniglia <ryan@readyset.io>
Reviewed-by: Marcus Bastian <marcus@readyset.io>