mir/rewrite: Recursively update columns of filters
It was pointed out in a review that when we update the columns list of
filter (or topk, latest, or identity) nodes due to having rewritten
their parents, we have to do so *recursively*, since there might be any
number of those nodes below the node we've just rewritten. This does
that by first generalizing the TopoAncestors iterator into a generic
Topo iterator that can go in either direction (via a type parameter, so
we get monomorphization), then using the new ability to iterate
descendants of a mir node topologically to move down the descendants of
the node we just updated and doing the necessary bookeeping on column
lists. This required (unfortunately) also moving things around to avoid
runtime borrow errors on refcells, which seems to be the running theme
of this whole function.

This *entire* thing would have been completely unnecessary if we had MIR
nodes calculate their column lists on-demand, rather than storing them
as a field - we should do that refactor as soon as possible.

Change-Id: I0bb01a32aaa3a4b2c8bef0d4b803cdb6dc8ff0de
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1993
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>