ops/benchmarks: migrate container image to new mariadb apt package name
Recently during CI benchmark runs, we've started to see failures during
benchmark container image builds, with an error message of:

Version '1:10.3.32-0ubuntu0.20.04.1' for 'mariadb-client' was not found

After further inspection, it looks like sometime after initial
development of the benchmarks, we may have patched this issue in the
readyset-mysql Dockerfile to mitigate this issue.

This CL aims to backport that change into the benchmarking Dockerfile.

Change-Id: Id6a34168862cb9023a1ffebe6f214cb657e14684
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1994
Reviewed-by: Justin Miron <justin@readyset.io>
Reviewed-by: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI