ops/cfn: allow VPC endpoints to be conditionally disabled (non-breaking)
The goal of this CL is to allow us to spin up the VPC supplemental stack
with the VPC endpoints toggled on or off as desired.

This change has been made in such a way that it should not break any
implementations of this stack which do not have this new parameter set.
The existing behavior has been retained.

This changeset will unblock provisioning of ReadySet cfn stacks in
existing VPCs. In these scenarios, we may not necessarily want the VPC
endpoints, depending on what automated process owns the source network.

Change-Id: I2b11d4b388d215c9c7e37f934f0af482ac2eb360
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1995
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Sasha Fox <sasha@readyset.io>