Add prep/exec functionality to DatabaseConnection
We have an abstraction around DatabaseConnection which allows us to
abstract the mysql or postgres client away to help de-dup testing. This
commit adds a prepare method, and a way by which the statement returned
by prepare (which in both the mysql and postgres case internally holds a
statement id) can be fed back into execute.

Now we will execute the passed in prepared statement if supplied, or if
a string is supplied, we will simply pass that through (and the
underlying client will both prepare and execute it).

This commit also fixes up some places in the runner where we could have
potentially errored out if the query we were testing was an internal
command of ours (which would not be interpreted by the underlying client
as a recognized Statement).

Change-Id: I3e075b9b550415f792c1f90615a687f551cb6ef9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1997
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>