ops/cfn: create benchmark CFN template that uses existing networking
RE: ENG-1124

The goal of this CL is to provide us with a usable MySQL super stack
that can be deployed into an existing AWS environment's VPC. The VPC
supplemental stack in this implementation will not create VPC endpoints.
The goal is to simply add the needed VPC security groups to the VPC and
then remove them after the tests complete.

RDS instances for benchmarking will be different from a standard RDS
instance we might deploy to a customer's environment. Benchmarking RDS
instances use a DB snapshot, which includes information such as the
username, password, etc.

So with this CL, for benchmarking, we should have 1 combined
MySQL/Postgres super stack, and 1x RDS instance stack which can handle
both MySQL and Postgres instances.

You may also notice that there is no Bastion host in this configuration.
The idea is for us to leverage Tailscale and internal AWS networking to
access the resource deployed in this stack vs needing to proxy all ops
through a Bastion host/tunnel.

Change-Id: I0168ca41d7ccc26c46dfe7306db4ea8c4882a5a2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2007
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>