noria-client: Calculate keys for page number
Implement the most basic form of the "client" part of pagination, which
involves dividing the provided parameter for OFFSET by the LIMIT value
in the query, and passing that as the value for the "page_number" key.

To accomplish this, this adds the actual value for the LIMIT as a field
to the corresponding ViewPlaceholder returned by noria-server, which has
been renamed from ViewPlaceholder::Offset to
ViewPlaceholder::PageNumber, to make it clearer that the actual *view*
wants a page number supplied, not an offset.

Currently offset values that are not integer multiples of the LIMIT are
unsupported - later we'll accomplish that by doing two lookups for
subsequent page numbers and taking the result set from those results in
combination (though this likely requires making multi-key lookups return
as a single result set, same as re-aggregating multi-key lookups does).

Refs: ENG-1050
Change-Id: Ib3a02305ec5ad67a169385494e7dcd251f0811b3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2003
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>