nom-sql: limit+offset only support Literal
As it turns out, the LIMIT and OFFSET clauses in a SQL query only
support *literal* values, not arbitrary expressions - so we shouldn't be
parsing arbitrary expressions there, as there's no world where we'd
support them.

This is a second attempt at bb30906b5def3d49c3dd28059f29d3580c536ba2,
which it seems like due to me mis-resolving merge conflicts didn't
actually apply the full change that I wanted it to.

Change-Id: I4a7bd51db929a82f520e82469343dc144a941e96
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2004
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>