noria-client: Autoparametrize OFFSET
Make the autoparametrization rewrite pass in noria-client also replace
literal OFFSETs with parameters, so they can be supplied as part of a
lookup key to a paginated view.

Getting this working required moving the extraction of limit columns to
after autoparametrization, to avoid sending the wrong number of
parameters back to the client for the query.

Note that I've elected *not* to gate this behind a feature flag for the
sake of runtime complexity - an autoparametrized OFFSET query sent to a
noria-server instance that has pagination disabled will just return
unsupported and go to fallback anyway, so I figured it was best to avoid
yet another runtime config flag.

Fixes: ENG-1135
Change-Id: Ibe27c46a8f7b046ec991082eb2453885d0d87831
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2006
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>