installer: Prompt for and pass SSM param for password
Rather than just prompting the user at the beginning for the SSM
parameter path for the password, then making them look up the KMS key
ARN that the parameter was encrypted with when actually creating the
stack, give the user the option to either input a password directly, or
select from a *list* of available SSM parameters, using the default SSM
KMS key in the former case and automatically determining the KMS key
that the parameter was encrypted with in the latter case. Then, when
creating the actual readyset cloudformation stack, pass both these
values directly as non-modifiable parameters, along with all other
required parameters. As a result of these changes, the installer script
now has no parameters in the readyset stack that it prompts the user for
and doesn't supply on its own.

Change-Id: I2f99241daa7ee2c98b86ee1f875829cc94182d45
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2011
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Jay Bensal <jay@readyset.io>