ops/terraform: new buildkite IAM role and module for k8s federation
Creating a new area for Terraform managed-services resources to live
(./ops/terraform/managed-services) as part of the benchmarking in CFN
intiative, instead of us putting all the new TF resources in the
Substrate area, it makes more sense to just lay the new foundation now.

As part of the initial foundation, I've added an IAM role, k8s role, and
relevant IAM policies to help Buildkite agents launch resources in k8s.

Additionally, since the IAM role created is relevant to the new
buildkite queue, I've included that in this CL as well. A separate queue
is recommended if we want to apply least priviledge. Basically the goal
is to NOT give every step in every build direct access to our k8s
infrastructure

Change-Id: I96d5ff2c7afb68c3d66ac81fa680e0fec71e8927
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2018
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>