keys: Return replay paths in replay path order
Previously, the various key-tracing algorithms in `controller::keys`
returned replay paths in *trace* order, meaning the *target* of the
replay paths are first, with the *source* of the keys coming last. This
was backwards from how replay paths are actually used in dataflow, where
the path represents the sequence of nodes a replay packet follows when
actually being executed, with the source node coming first and the
target node coming last - to make this work, we had to `reverse` all the
replay paths when actually constructing them in materialization::Plan.
This was vaguely excusable when the actual API was
`keys::provenance_of`, but now that the functions in `controller::keys`
have the word "replay_paths" in their name, this makes very little
sense.

To make things easier to understand and more consistent, this commit
changes the functions with "replay_paths" in their names in
`controller::keys` to instead return paths in *replay order*, iterating
over them in reverse order at the call site if we need to (which only
actually happens once).

Refs: ENG-1120
Change-Id: Ie8bc3bfa0591a5c4616628689e7e34e2dbd7cd48
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2022
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>