cfn: Fix default port number for MySQL template
Debugging an issue with the installer looked like there was a
a connectivity issue. In reality, if the adapter or server cannot
connect to the upstream database, the connection hangs. (This should be
fixed)

After digging around, I finally SSHed into one of the machines and found
that the port number that was being used was wrong and was instead 5432.

The installer is not passing the port number through unlike in the case
of the super stack which does. And the default port number of 5432 was
breaking it since it was trying to connect to the Postgres port.

This just encourages us more towards getting rid of the super stack and
our scripts and moving towards the installer as the one interface for
this given that we will be sharing the installer with our customers.

Change-Id: Ib999e54eaab9dd100c909bb3afa765fb9fc1e3d5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2028
Tested-by: Buildkite CI
Reviewed-by: Sasha Fox <sasha@readyset.io>