domain: Log replay paths in a more friendly way
Rather than using the derived Debug impl when logging the segments of a
replay path during trace-level logging in domain, which is super verbose
and hard to follow, define a lazy Display-wrapper around the segments of
the replay paths which logs them in a much easier-to-parse way. This
makes reading trace log output during debugging of replay-path-related
code way easier.

Change-Id: I536d2814561bd6342e2d6f3d43d656e75181bc01
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2031
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>