ops/benchmark: cfn automated benchmarking with k8s benchmark runner
Included in this CL are the remaining components necessary to run
automated benchmarks against a provisioned ReadySet Cloudformation stack
using Kubernetes.

Adjustments to the existing benchmark pipeline were needed in order for
us to start freshly building the benchmark runner container image, which
is needed for running things in k8s.

Adjustments have been made to the existing benchmarking script so that
we can pass the results files back through STDOUT to the Buildkite
agent. The idea here was to avoid duplicating the existing script so
that we may continue to maintain a single script for all non-local
benchmark executions. This might not work forever, but for the time
being it seems like a good way to go. #DRY

A mini-Helm chart is included in this CL which effectively represents
the "benchmark runner" we launch in k8s to perform the actual
benchmarks.

Kubernetes node affinities and pod anti-affinities are used to ensure no
two runner pods are scheduled on the same EC2/EKS worker node at any
given moment. In an effort to optimize the process for consistency, the
infrastructure which underpins the benchmark runner pods is locked to a
single AZ.

Change-Id: I95692ab4ff75163463af6cc9bf9749452a980265
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2032
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Sasha Fox <sasha@readyset.io>
Terraform-Plan-Review: Sasha Fox <sasha@readyset.io>