noria/server: ReadRequestHandler: Replace nested Either with custom Future enum
This recovers the ~1% performance penalty that had been imposed on
SELECT queries by adding tracing.

Previously, this had explicit enum variants for instrumented versions of
each.  Those are no longer necessary, so this isn't as substantial of an
optimization as it used to be, but it does still offer a small win.

Change-Id: I0dc0ee09319849fd20373209b89be5519e7e7f80
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2039
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>