docker/monitoring/docker-compose.yml:  + jaeger
Jaeger is used for monitoring traces/spans.

By default, it stores spans in memory; as configured, it uses ephemeral
on-disk storage with an embeded database.  Neither of these are
production-capable configurations; for production, either Cassandra,
Elasticsearch, or InfluxDB is required.  This is a workable
configuration for CI and developer use.

Change-Id: Id7673d56392f5a7e54b6cbedfd6b3ffa61a6bd88
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2040
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>