dataflow: validate key length in SingleState::mark_filled
When modifying code that deals with the creation and manipulation of
replay paths, it's really easy to accidentally introduce bugs that cause
tags for replay paths to point at the wrong indices in state - which can
result in, among other things, attempting to perform writes into indices
in state with keys of the wrong length. When this happens, it's way
better to get an error message indicating the actual and expected key
length rather than just an unwrap error - to that end, this change
introduces some asserts to `SingleState::mark_filled` that check the
length of the key before actually trying to build the tuple index into
the map. This would already panic at an unwrap, but now it does so with
a better error message!

Change-Id: Ie6c415af3cdea0f9829bfa1411a7c4e50ac27b62
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2043
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>