noria-client: Log during connection establishment
When a user connects to a readyset adapter, the first thing we do is
establish the connection to the upstream database for passthrough and
fallback queries. If this upstream connection fails or hangs, the user
would previously get no feedback that anything at all was happening,
since we didn't log that we'd accepted a connection until *after* the
upstream connection was successfully established!

To aid in debugging those sorts of issues (which are exceedingly common
in AWS land what with security groups being the pit of despair that they
are), this commit moves the "Accepted new connection" log message *as
soon* as the TCP socket is established, before we actually start trying
to do anything with it, and also adds some logs for the actual upstream
db connection establishment process for MySQL mirroring what we already
had for PostgreSQL, including the host, port, and username of the db
we're trying to connect to.

Change-Id: Ie01d16a6bee1b396e320d9bf24d7f7dfc549aa1c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2047
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>