adapter: Log + return upstream connection errors
Previously, any errors that occurred when connecting to the upstream
db (connection refused, authentication errors, etc.) would just get
unwrapped from a Result, dropping the connection and causing the client
to fail with an opaque protocol-level error message (since we never sent
the handshake packet). Now, we actually log those errors, and also
return them to the clients using their native protocol error packets, so
that both mysql and postgresql clients will actually log the error that
we encountered connecting to the upstream before exiting.

To accomplish this, this adds a new "immediate_error" method to
noria_client_adapter::ConnectionHandler, used to return errors
encountered immediately on connection establishment before we have a
backend to run on - for both mysql and psql this calls new top-level
functions to return these immediate error packets. In psql-srv in
particular this is a *little* messy, since we still need to thread
through a type parameter for the Backend (since *certain* error messages
can care about the types of rows and values returned), but it's not too
bad and it's pretty isolated.

Change-Id: I1f39e34045e1120623d521cb6f93d1afe2fbaa0a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2048
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>