Reuse select queries
Whenever a new select query is added we first check that
it does not yet exists under a different name, and if it
does we simpy alias it to the existing query. This allows
us to avoid creating spurious readers, when they already
exist.
In addition, since we rely on Hash to produce a unique
query id, but  Hash only generates a 64bit digest, which
has a pretty high chance of collision, and it is something
that we can't afford, this change is using 128bit digests
for query ids.

Change-Id: Id83c334b7b7912c75233d4408953e4d1db97c00a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2054
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>