adapter: Don't use an Arc for QueryStatusCache
Every noria-client-adapter binary (noria-mysql or noria-psql) uses a
single global QueryStatusCache value that it shares across all
connections, and lives for the lifetime of the process. This is a
textbook usecase for using Box::leak to get a &'static value over
wrapping things in an Arc - there's no need to track reference counts
like this since we know the only thing that'll ever free this memory is
terminating the process! This has a very small, but present, performance
benefit too since it avoids the atomic increment and decrement on
creation of new connections.

Change-Id: I09354763ade56b078bb9ad44aa21618b2a862f85
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2053
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>