mir: Rewrote MIR serde algorithm.
The MIR serde algorithm was not working properly, as it sometimes left the MIR
state inconsistent.
This commit replaces the previous serde algorithm for a more sturdy one, using
`petgraph::Graph` as the underlying, serializable structure.

Given the current state of MIR, since the references to the nodes can be spread
across multiple MIR queries (aka, MIR graphs), we had to create custom version
of the structures that hold the `MirNodeRef` references, by almost copy-pasting
the same structures into ones that only held "soft" references (such as node
indices from `petgraph::Graph` or just an identifier for the nodes).
That, although awful, was not a big deal for the main structures
(`SqlIncorporator`, `MirQuery` and `SqlToMirConverter`).
However, the only structure that could not be adapted is the `MirNodeInner`,
specifically the `MirNodeInner::Reuse` variant, which holds a reference.
Since (at the time this commit was written) the reuse feature is not working, we
are not considering that case, and we are therefore disallowing the use of
reuse with this (since using it would cause a stack overflow during the serdes
computation for `SqlIncorporator`, when persisting the dataflow state to the
authority).
Therefore, we've removed the `enable_reuse` flag from the server.
We can re-enable reuse after we refactor MIR and move away from a distributed
graph (based on node references) into a centralized graph (like
`petgraph::StableGraph`).

Refs: ENG-1161
Change-Id: Icf5e0e8be8739465ce14bea539c4f24a73d4903c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2061
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>