dataflow: Temporarily disable remapping for straddled joins
Getting self-generated columns in nodes working, like we want to do to
make pagination partial per group or to optimize aggregates to
pre-calculate the aggregate at the target of the replay, requires a
redesign to how replays work for nodes which generate columns from
*their parents* such as straddled joins. Right now, optimizing
aggregates to pre-calculate at the target of the replay is pretty high
priority, so for the time being this commit just makes any straddled key
on a join return `ColumnSource::RequiresFullReplay`, which will force
that node to be fully materialized. In a subsequent change, I'm going to
make the materialization planner return an error for any
ColumnSource::GeneratedFromColumns where the node is not the same node
being traced through - once we decide we want to get partial straddled
joins working again we can revert this commit and remove that error.

Change-Id: Ibae7c7b01bfdb41d9981914def6bfb29f604c418
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2063
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>