dataflow: Allow nodes to generate columns from themselves
This commit makes a number of changes to the way remapped
upqueries (represented by the `ColumnSource::GeneratedFromColumns`  enum
variant and the `Ingredient::handle_upquery` trait method) work, both in
the materialization planner and in dataflow execution, to allow nodes to
conceptually generate columns "from themselves" - meaning a replay
request to some set of generated columns on that node will be translated
to an upstream replay targeting a *different set of columns on the same
node*.

The crux of a lot of what needed to change here to get this to work
stemmed from the original design of remapped upqueries, which were
really only built to support straddled joins. Essentially, the way those
worked was that the join would return `GeneratedFromColumns` with the
subset of the column indices in the straddled key that came from each
parent, *pointing at the parent* - the replay path generation code would
then generate replay paths going through each of those parents, then to
the join, but with the target index being the actual straddled key
inside of the join itself! For example, if we had a join n3 of nodes n1
and n2 keyed on columns [1, 2], where column 1 refers to column 0 in the
left and column 2 refers to column 1 in the right, we'd get a replay
path from n1[0] -> n3[1, 2] and n2[1] -> n3[1, 2] - note the change in
key column length along the replay paths! When executing those replays
- which would be keyed only on the subset of the original replay key
that actually maps to the individual parent, not the whole straddled
replay key - we had to have an extra conditional at the target of the
replay to avoid attempting to mark a two-column hole filled with a
one-column key value. This conditional, however, would prevent us *ever*
filling the hole for a node that generated columns from itself, as no
segment of the replay path would actually contain the remapped lookup
key for the node.

Now, instead of doing that, we have the replay path stop at the
actual *hole* that's being filled - aka the actual IndexRef returned by
column_source as part of the GeneratedFromColumns enum variant, and
remove the conditional that prevents that hole from being filled. Since
we can satisfy lookups for a larger set of columns with lookups into a
smaller set of columns followed by a linear scan, in the case of nodes
that generate columns from themselves it's sufficient to simply mark
that target hole as filled, then re-try the original lookup. We'll have
to do a little extra work once we want to bring back nodes that generate
columns from their parents (of which straddled joins are currently the
only example), as with the current way things work we'd just replay to
that parent and stop there, and the join itself would never see the
rows. But that's a problem for future us.

A couple other miscellanious things that've changed here, either
incidental cleanup or to help this happen:

- ColumnMiss now works with LocalNodeIndex instead of global
  NodeIndex (requiring it to inline the fields from ColumnRef rather
  than using it directly) - not only does this avoid a whole lot of
  translation from LocalNodeIndex to global NodeIndex both inside of the
  node and the domain, it also allows us to...
- Get rid of raw_replay_paths in the domain, which was always a bit of a
  wart and wasn't super easy to explain - but now it's no longer
  necessary for anything!
- We now remap upqueries *before* calling `find_tags_and_replay`, in
  `on_replay_misses`, rather than *inside* `find_tags_and_replay` - this
  is mostly in service of inserting the correct keys inside of Waiting,
  since the actual target of the upstream replay is now different than
  the lookup index of the downstream replay. The only other place this
  is called is in the direct reader replay request code - we don't need
  to worry about remapping upqueries there, since readers can't generate
  columns.
- We avoid a lot of extra twiddling and complexity in materialization
  planning by having the replay path generation code just stop the path
  at generated columns and say that it did so via a boolean field
  returned in a new RawReplayPath struct, rather than returning two
  paths - this works because materializations are calculated for *all*
  new nodes in the graph in *reverse topological order* - we just add a
  replay obligation for that set of columns at the node, then generate
  the upstream replay path for the remapped columns once we get to it.
  This is in fact much closer to how the materialization generation code
  already works for regular, non-remapped materializations.
- Because we no longer need to do any key remapping at the target of an
  upstream remapped replay, we can remove the `active_remaps` field from
  Domain, and a lot of code that only existed to support it.

Finally, to prove that this all works as expected, this commit goes
ahead and just implements column_source and handle_upquery for the
Paginate operator, making it partial per-group - the existing
integration test for that node covers this.

Fixes: ENG-1120, ENG-1106
Change-Id: I7aa6387e8a562746a0b04a83b41554825af165a7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2066
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>